Prototype of peaklets-level (x, y) S2 position reconstruction (#1482)
* Prototype of peaklets-level pos-rec

* Simplify dependency tree

* Add TODO

* Pos-rec for merged_s2s

* Recover all possible data_types

* No 'NT'

* Debug

* Restructure the dependency tree with straxen style

* Still keep CNN and GCN

* Debug

* Debug

* Debug

* Debug

* Renaming

* Debug

* Remove CNN and GCN

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* A minor change

---------
