+++ straxen/plugins/events/_event_s1_positions_base.py
@@ -20,7 +20,6 @@ class EventS1PositionBase(strax.Plugin):
 
     algorithm: Optional[str] = None
     compressor = "zstd"
-    parallel = True
 
     min_s1_area_s1_posrec = straxen.URLConfig(
         help="Skip reconstruction if area (PE) is less than this",
