+++ straxen/plugins/events/s2_recon_pos_diff.py
@@ -15,7 +15,6 @@ class S2ReconPosDiff(strax.Plugin):
 
     __version__ = "0.0.3"
 
-    parallel = True
     depends_on = "event_basics"
     provides = "s2_recon_pos_diff"
     save_when = strax.SaveWhen.EXPLICIT
