+++ straxen/plugins/merged_s2s/__init__.py
@@ -1,2 +1,8 @@
 from . import merged_s2s
 from .merged_s2s import *
+
+from . import merged_s2_positions_cnf
+from .merged_s2_positions_cnf import *
+
+from . import merged_s2_positions_mlp
+from .merged_s2_positions_mlp import *
