+++ straxen/plugins/peaklets/__init__.py
@@ -6,3 +6,9 @@ from .peaklet_classification import *
 
 from . import peaklet_classification_som
 from .peaklet_classification_som import *
+
+from . import peaklet_positions_cnf
+from .peaklet_positions_cnf import *
+
+from . import peaklet_positions_mlp
+from .peaklet_positions_mlp import *
