+++ straxen/plugins/peaks/__init__.py
@@ -13,14 +13,14 @@ from .peak_classification_bayes import *
 from . import peak_s1_positions_cnn
 from .peak_s1_positions_cnn import *
 
-from . import peak_positions
-from .peak_positions import *
-
 from . import peak_positions_cnf
 from .peak_positions_cnf import *
 
 from . import peak_positions_mlp
 from .peak_positions_mlp import *
+
+from . import peak_positions
+from .peak_positions import *
 
 from . import peak_proximity
 from .peak_proximity import *
