+++ straxen/plugins/peaks/peak_basics.py
@@ -16,7 +16,6 @@ class PeakBasics(strax.Plugin):
     """
 
     __version__ = "0.1.4"
-    parallel = True
     depends_on = "peaks"
     provides = "peak_basics"
