+++ straxen/plugins/peaks/peaks.py
@@ -1,3 +1,4 @@
+from typing import Tuple, Union
 import numpy as np
 import strax
 import straxen
@@ -18,10 +19,9 @@ class Peaks(strax.Plugin):
 
     __version__ = "0.1.2"
 
-    depends_on = ("peaklets", "peaklet_classification", "merged_s2s")
+    depends_on: Union[Tuple[str, ...], str] = ("peaklets", "peaklet_classification", "merged_s2s")
     data_kind = "peaks"
     provides = "peaks"
-    parallel = True
     compressor = "zstd"
     save_when = strax.SaveWhen.EXPLICIT
 
@@ -66,6 +66,6 @@ class Peaks(strax.Plugin):
                 peaks["time"][to_check][1:] >= strax.endtime(peaks)[to_check][:-1]
             ), "Peaks not disjoint"
 
-        result = np.zeros(len(peaks), self.dtype)
-        strax.copy_to_buffer(peaks, result, "_copy_requested_peak_fields")
+        result = np.zeros(len(peaks), dtype=self.dtype)
+        strax.copy_to_buffer(peaks, result, f"_copy_requested_{self.provides[0]}_fields")
         return result
