+++ tests/plugins/posrec_plugins.py
@@ -1,16 +1,16 @@
 """Run with python tests/plugins/posrec_processing.py."""
 
 import os
-import strax
-import straxen
-from _core import PluginTestAccumulator, run_pytest_from_main
 import numpy as np
+import strax
+from _core import PluginTestAccumulator, run_pytest_from_main
+from straxen.plugins.peaklets._peaklet_positions_base import PeakletPositionsBase
 
 
 @PluginTestAccumulator.register("test_posrec_set_path")
 def test_posrec_set_path(
     self,
-    target="peak_positions_mlp",
+    target="peaklet_positions_mlp",
     config_name="tf_model_mlp",
     field="x_mlp",
 ):
@@ -42,7 +42,7 @@ def test_posrec_set_path(
 @PluginTestAccumulator.register("test_posrec_set_to_none")
 def test_posrec_set_to_none(
     self,
-    target="peak_positions_mlp",
+    target="peaklet_positions_mlp",
     config_name="tf_model_mlp",
     field="x_mlp",
 ):
@@ -56,7 +56,7 @@ def test_posrec_set_to_none(
 @PluginTestAccumulator.register("test_posrec_bad_configs_raising_errors")
 def test_posrec_bad_configs_raising_errors(
     self,
-    target="peak_positions_mlp",
+    target="peaklet_positions_mlp",
     config_name="tf_model_mlp",
 ):
     """Test that we get the right errors when we set invalid options."""
@@ -73,8 +73,8 @@ def test_posrec_bad_configs_raising_errors(
     with self.assertRaises(FileNotFoundError):
         plugin.get_tf_model()
 
-    dummy_st.register(straxen.plugins.peak_positions_cnf.PeakPositionsBaseNT)
-    plugin_name = strax.camel_to_snake("PeakPositionsBaseNT")
+    dummy_st.register(PeakletPositionsBase)
+    plugin_name = strax.camel_to_snake(PeakletPositionsBase.__name__)
     with self.assertRaises(NotImplementedError):
         dummy_st.get_single_plugin(self.run_id, plugin_name)
